server/requirements.txt
astroid==2.4.2
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
colorama==0.4.3
connexion==2.7.0
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
inflection==0.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
openapi-spec-validator==0.2.8
protobuf==3.12.2
pyasn1==0.4.8
pylint==2.5.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jose==3.2.0
PyYAML==5.3.1
requests==2.24.0
rope==0.17.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
swagger-ui-bundle==0.0.8
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
